Fixes APY estimation
There was a problem with the `Staking.currentAPY` calculation.

The initial bug report was the following:

```
The function currently returns the APY for the following period of time: (now, now + 1 day). I’m not sure if that was the intention, or  (now - 1 day, now + 1 day).

Before the start date it should either return 0/revert. Or it should just return the first day’s APY. But it’s always starting from uint currentReward = calculateReward(startDate, current, amount); which gives you unpredicted result. Because you expect 1-day reward, but startDate - current will be lower than 1 day and always changing.
```

Since this function was added at last minute, and is purely for
informative purposes only (to display an estimated APY on the staking
UI, with no actual consequences to the staking calculations themselves),
its full behaviour was not properly thought out

After discussing, the desired output is:

* Before `startDate`, it should have the same result as if it were
  called exactly at `startDate`. So rather than returning `0`, we must
  show what the estimated APY will be once it starts

* The range (now, now + 1 day) is correct, as we want to provide an
  answer to "what will be my APY should I stake right now?"

With that in mind, the function was updated, and corresponding tests
were refactored to better clarify the expected results